backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langgraph==0.2.28
langchain-core==0.2.39
requests==2.31.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
arxiv==2.1.0
pydantic==2.9.0
boto3==1.35.0